LayoutTests/fast/viewport/viewport-warnings-4.html

<head>
    <title>Truncated value.</title>
    <meta name="viewport" content="width=123x456">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(160, 320, 480, 320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
